Wire up a builtin firmware mechanism.
This uses the libc init/ctors stuff to allow module firmwares to register
themselves when they've been compiled and linked into the main firmware.

So, once they're registered in the registry we can lookup the metadata for the
module based on the header tuple (manuf, kind, version)

Eventually these should be separate binaries.