Split module scanning and lookup and tie into readings taking.
This also breaks readings taking into a separate class, moving stuff outside of
the task itself for easier testing. I'm... a little unsure about this design but
this should be ok for the near term until I can think up a better way. I think
this is just over complicated but the memory allocation going on here kind of
steered me towards this design.