Power cycle modules and hibernate if readings fail.
This came about as a result of some quirks with the RTD sensor. I'm
thinking of asking Atlas about this stuff. Basically the sensor
wouldn't give a reading at startup, even though the wake command worked.

So, we power cycle and also sleep if readings fail so the 2nd wake is
actually correct. I tried a loop here, no dice. Just fussy, but can wait.