Fix seek error that would find the wrong record number.
The crux of this was the seek would find a valid tail location in a later block
from the writing of longer/larger file. This would locate the precise tail of
the other file and so no further seeking would occur within the block and the
record was being left as 0.