Use instrumention functions to verify GOT value.
The main tricky thing here is the use of an absolute address to where the
expected GOT value is. Our fkb loader will push the given GOT value onto the
stack and so that's where we look. If we try and use a static variable for this
then that ends up getting relocated and depending on R9 to begin with!

I'm hoping this helps narrow down places where R9 is being clobbered.