Add option to always leave GPS on. FK-1810
This works the same way the WiFI always on option is. Basically sets
an insane duration and treats that as always on mode. It will also
disable turning off GPS after having a fix for the configured required
fix duration.

Another fix is to only check GPS status after parsing pending
characters, which should be faster.