Faster seek when finding the end.
If we're finding the end we can can skip whole blocks during the
find_blocks_after, otherwise we check each block. Eventually, when
finding a particular record we can look at the records to tell if we
need to scan them, this way optimizes the end seek and right now the
record seek will just be slower until we get to that.