Chasing a truncated file upload bug. FK-2300
I suddenly started seeing truncated uploads from our long running
station. Oddly after a firmware upgrade, even though this appears to
have been an issue before. I even reverted to make sure.

The bug was that with a bad block that makes the binary search go
backwards when we should have gone forwards, we need to do a linear
scan. The bytes in the blocks that were scanned weren't being properly
accounted for and so the position would be wrong.

This means when the position of the end of the file was wrong, so when
we would seek to the end and then subtract to get the upload size, we
weren't uploading the right number of records, and were actually
cutting off the middle of the last one.