Improve error handling on firmware uploads.
This is kind of annoying..... some HTTP clients will fail with a
Broken Pipe/EPIPE when we immediately write back an error w/o reading
the entire request. So this is a work around for that, we just read
the entire request into /dev/null and then reply with the fancy error.