Trial mode of a deep sleep during very dormant periods.
When the scheduler goes from having workers to having none, we ask the
scheduler when the next task will need to run.

If that time is greater than our minimum deep sleep then we check to
see if a deep sleep is possible, which means: no networking, the
network, display and gps tasks are all stopped.

Otherwise, we deep sleep.

This means we'll only do the one sleep after the worker stops, for now.