Firmware for unmetered weather stations.
These are basically standard weather setups w/o the metering support on the sensor board. We're lucky the unmetered sensors are the first in the payload, because the approach this is taking is to just truncate the readings payload from those. I needed to maintain compatibility in the core firmware though so what we do is the the unmetered fw will flag reveal itself by using bogus magic values in the rain/wind. Not the best solution, but certainly better than having a separate module? Maybe.