New approach to rolling rain and wind readings.
So this is a little embarassing. Before the major change to
aggregated I was polling the weather information every minute,
regardless of the reading interval the user chose, which
led to pretty smooth data.

With the move to the new appraoch to the weather data, the information
piles up on the SAMD09 and so the accumulated numbers make no
sense.

This changes modifies the default rain and wind readings to be more
sensible. Rain is inches/hour and rolling. We no longer reset at the
bottom of every hour.

Wind is a 30s average, calculated from the 120s of wind data we keep.