More open options, including Append which we use in CSV.
Saw an odd scenario where the meta record looped back, and I believe
that this had something to do with having to rescan while reading the
file or something. Not sure yet, but I'm stuck until I can reproduce
it on a station that also isn't experiencing some other issues related
to flash memory.