Replace atlas status with module configuration in EEPROM.
This also strips some of the calibration code away and instead models
the event as us writing and reading serialized ModuleConfiguration's
to the EEPROM memory. Module statuses are now read from this EEPROM
memory as well. Instead of starting with them as per-module, I'm
trying to keep as much calibration logic generalizable as I can.