Juggle in memory module configuration.
This is wasteful, memory wise and could lead to issues if people do
lots of calibrating at once. It's a workable implementation until we
find a better way to juggle module memory.

What our problem here is, we allocate the configuration in initialize
from the module factory pool and then need to ensure that the
configuration after API calls has the same lifetime.