Force GCC to compile access to external fns via GOT.
GCC was making PLTs that weren't using R9 to access the
GOT. Supposedly there is a flag -fno-plt that forces those accesses to
go through the GOT, though that's not implemented.

The workaround is to just use function pointers, which get looked up
via the GOT in the usual manner for data.

https://answers.launchpad.net/gcc-arm-embedded/+question/669758