First working dynamically loaded module!
Still quite a ways to go, but this is a proof of concept at
least. Right now this compiles the dynamic test module and then dumps
the binary to an include that gets linked in. Then, there's a special
worker wrapper - Process, that will run that dynamic code after
preparing the memory space. That preparation is mostly allocating
memory to symbols and filling in the GOT table for the process,
linking to the operating system functions.

A major difference with this one, is that the logging call happens in
the task process rather than being a syscall.

Notice also that the code has to be dword/u32 aligned! If it's just
2-byte aligned things may still sort of work at frist, but calculating
pc based offsets will be a mess.