feat: Proving some workings of the LFS storage.
We'll be building the longer term storage mechanism on top of the
POSIX file api that littlefs provides. This will require some special
consideration but should be far more stable and reliable than the
existing flash code.

In order to support freeing up space, we need to divide the data up
into files so that we can remove older files to claim that space. To
facilitate this we begin by splitting the data up across files based
on some heuristic that tries to right size the files in memory.

We can then use the very handy custom attributes to overlay the block
number mapping on top of those files so they can be read as one file.