fix: Refresh directory after file is fully written.
We know we need to do this, but no reason to do this before we've written the file, as that'll make the map old already. So we do this after we've written the record. This also helps because the directory listing will see the fully formed file, attributes and all.

I still worry there's a situation where the attributes aren't there cause lfs wasn't able to fully recover.