fix: Fix serious memory bug. Malloc page race. FK-3589
The actual updating of the available flag was still happening outside
of the disabled IRQ section, so there were still serious races in the
page allocation.

I've also moved the memory garble/zero outside of the IRQ. We do this
before now, while we still own the page and are sure to make sure we
have a valid page before we do so.

Conflicts:
	fk/memory.cpp