hack: Force original fixed page size.
True Koxia page size is 4096, only the earlier code would always
assume a page size of 2048. So, we can't assumed a fixed geometry
based on the chip we have. This will be the case until we can get
ourselves off the old file system.  So this code starts by just
assuming the old page size and then when we learn that the choice
is up to us we can use the proper one.