feat: Add low power and dangerously lower power status.
We'll consider ourselves in low power startup if the battery voltage
is < 3.7V and the battery gauge is available. In this mode we'll skip
starting high power services like WiFI/radio/screen unless the user
presses a button.

Dangerously low power is set when the INA219 is inaccessible or
battery voltage < 3.5V and > 0.3V. This behaves the same as low power
mode except we also disable taking readings.