feat: Alternative timeout for distance readings.
This will attempt to read distances for up to a given timeout, rather
than trying to read N of them. It seems like in some situations
there's a longer lag to the sensor starting up and so we get it's
startup banner and never get a reading. This should read further past
that banner.