fix: Increased log verbosity causes networking instability.
Hello future programmer, you may be wondering why this is here and
even be tempted to remove this so you can chase a problem. You will
want to be very careful about this. This is here, because if we log
heavily during periods of high logging activity the WiFi module starts
to experience all kinds of truama. My theory is that the logging gets
in the way of the IRQ handling. I believe you can even exacerbate this
by moving your AP further away so that latencies are higher and IRQs
happen during periods of more intense logging, say when accessing the
file system.

You've been warned.