fix: Ignore bad schedules when configured.
Before we'd just take whatever was given, this does a better job
excluding scenarios that are obviously invalid. For example, ignoring
network scheduels with a zero duration. It then applies this reasoning
to the others.

I'd like to remove the fallback stuff eventually, but until we're only
ever sending complex schedules we can't.