fix: Check all viable record addresses after failed header read.
I was looking at logs from a corrupted station and was taking the time
to actually look over the binary dump and comparing logs to the data and
noticed something odd.

Before, we weren't logging the address we were reading the record
header of, after I added that in, a potential issue stood out:

00000184 verbose storage: [0] 0x00780200 reading head
00000184 verbose storage: [0] 0x00780200 valid head
00000184 verbose storage: [0] 0x00780200 seeking  293/ 221 (R-14733)
00000184 verbose storage: [0] 0x00780325 reading head
00000184 verbose storage: [0] 0x00780325 invalid head
00000184 verbose storage: [0] 0x00780400 seeking  293/ 221 (R-14734)
00000184 verbose storage: [0] 0x00780525 reading head
00000184 verbose storage: [0] 0x00780525 invalid head
00000184 verbose storage: [0] 0x00780600 seeking  293/ 221 (R-14735)
00000184 verbose storage: [0] 0x00780725 reading head
00000184 verbose storage: [0] 0x00780725 invalid head
00000184 verbose storage: [0] 0x00780800 seeking  293/ 221 (R-14736)
00000184 verbose storage: [0] 0x00780925 reading head
00000184 verbose storage: [0] 0x00780925 invalid head
00000184 verbose storage: [0] 0x00780a00 seeking  293/ 221 (R-14737)
00000184 verbose storage: [0] 0x00780b25 reading head
00000184 verbose storage: [0] 0x00780b25 invalid head
00000184 verbose storage: [0] 0x00780c00 seeking  293/ 221 (R-14738)
00000184 verbose storage: [0] 0x00780d25 reading head
00000184 verbose storage: [0] 0x00780d25 invalid head
00000184 verbose storage: [0] 0x00780e00 seeking  293/ 221 (R-14739)
00000184 verbose storage: [0] 0x00780f25 reading head
00000184 verbose storage: [0] 0x00780f25 invalid head

A few things are going wrong here. First, 0x00780200 is a valid
record, and then a valid record is never found again even though the
record number is increasing on the right hand side of those logs. Before,
this was impossible to see. Basically this is reusing the same record's
size each pass after the invalid record because we only continue in the
inner loop.

So, now we continue in both scenarios, which causes 0x00780400 to
actually get read for a header. Notice in the previous log we never
read that address, which is the correct one.