fix: Use our original sleep/wake code instead of TTNs.
As I was testing some incoming hardware I noticed that a few of the modules had
problems returning from sleep. After trying a dozen things I decided to try the
older sleep/wake code I had written for the rn2903 and lo and behold that
worked. With the increased timeouts, though I noticed things taking forever.
There was an extra read_line_sync in the wake function, which should only be
there if something is being sloppy. I removed that.