feat: Store record number on immutable records.
Did my best to avoid having to do this but I think it's the best solution. The
way we avoid having issues with the hash is to basically pass the hash that was
calculated with a 0 in place of this value, then when we decide that it's
changed we assign the value.