fix: Increase task priority during excited readings take.
We suspect that a context switch is causing longer pulses during EC readings
than we'd like, so this will bump the priority of the active task in that
situation.

It's not feasible to disable pre-emption completely here, at least with how
that's implemented because we do make service calls as part of the readings
process. I hope that increasing the priority is enough.