feat: Increase maximum configuration size.
This is used to size the sub pool we use to store the module's
configuration. Right now, it effectively sets an upper bound on the
total configuration size because we allocate and keep the serialized
copy around.

Ideally, we'd serialize that on demand. For reference, even omni's
configuration should never be more than 512 bytes.